update import DB

+++ modules/backend-dossiermgt/backend-dossiermgt-service/src/main/java/org/opencps/dossiermgt/action/ServiceProcessActions.java
@@ -91,7 +91,7 @@ public interface ServiceProcessActions {
 			String roleCode, String roleName, boolean moderator, String condition, ServiceContext serviceContext);
 
 	public long updateProcessStepDB(long userId, long groupId, long serviceProcessId, String stepCode, String stepName,
-			String sequenceNo, String groupName, String dossierStatus, String dossierSubStatus, Integer durationCount,
+			String sequenceNo, String groupName, String dossierStatus, String dossierSubStatus, Double durationCount,
 			String instructionNote, String briefNote, String roleAsStep, Integer checkInput,
 			ServiceContext serviceContext) throws PortalException;
